+++ pages/api/projects/[id].js
@@ -2,12 +2,13 @@ import * as fs from 'fs';
 import dbConnect from '../../../utils/dbConnect';
 import Projects from '../../../models/Projects';
 import { muterUpload, nextConnectonFunction } from '../../../utils/functions/apiHelper';
+import Admins from '../../../models/Admins';
 
 dbConnect();
 
 const apiRoute = nextConnectonFunction();
 
-const uploadMiddleware = muterUpload(2, 'projects').array('images');
+const uploadMiddleware = muterUpload(2, 'projects').fields([{ name: 'cover', maxCount: 1 }, { name: 'image', maxCount: 1 }]);
 
 apiRoute.use(uploadMiddleware);
 
@@ -23,26 +24,36 @@ apiRoute.put(async (req, res) => {
     const description = req.body.description;
     const conclusion = req.body.conclusion;
     const socials = req.body.socials.split(',');
-    const images = req.files.map(e => e.filename);
+    const url = req.body.url;
+    const onMain = req.body.onMain;
+    const date = req.body.date;
+    const cover = req.files.cover ? req.files.cover[0].filename : req.body.cover;
+    const image = req.files.image ? req.files.image[0].filename : req.body.image;
+
+    const isAdmin = await Admins.find({ 'token': req.headers.authorization }).populate('token');
+    if (isAdmin.length <= 0) {
+        req.files.cover && fs.unlinkSync(`./public/uploads/projects/${cover}`);
+        req.files.image && fs.unlinkSync(`./public/uploads/projects/${image}`);
+        return res.status(400).json({ statusCode: 400, message: 'Вы не авторизованны' });
+    }
 
     // get old images array of names
     const projectById = await Projects.findById(id);
     // if new images uploaded
-    if (images.length > 0) {
-        // delete old images
-        for (let i = 0; i < projectById.images.length; i++) {
-            fs.unlinkSync(`./public/uploads/projects/${projectById.images[i]}`);
-        }
+    if (req.files.cover) {
+        fs.unlinkSync(`./public/uploads/projects/${projectById.cover}`);
+    }
+    if (req.files.image) {
+        fs.unlinkSync(`./public/uploads/projects/${projectById.image}`);
     }
 
     const body = {
-        type,
-        title,
-        description,
-        conclusion,
-        socials,
-        images: images.length > 0 ? images : projectById.images,
+        type, title,
+        description, conclusion,
+        socials, cover, image,
+        url, onMain, date
     }
+
     try {
         const project = await Projects.findByIdAndUpdate(id, body, {
             new: true,
@@ -50,13 +61,13 @@ apiRoute.put(async (req, res) => {
         })
 
         if (!project) {
-            return res.status(400).json({ statusCode: 400 });
+            return res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
         }
 
         res.status(200).json({ statusCode: 200, data: project })
 
     } catch (error) {
-        res.status(400).json({ statusCode: 400 });
+        res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
     }
 })
 
@@ -70,13 +81,13 @@ apiRoute.get(async (req, res) => {
         const project = await Projects.findById(id);
 
         if (!project) {
-            return res.status(400).json({ statusCode: 400 });
+            return res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
         }
 
         res.status(200).json({ statusCode: 200, data: project });
 
     } catch (error) {
-        res.status(400).json({ statusCode: 400 });
+        res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
     }
 })
 
@@ -85,22 +96,27 @@ apiRoute.delete(async (req, res) => {
         query: { id }
     } = req;
 
-    const projectById = await Projects.findById(id);
-    for (let i = 0; i < projectById.images.length; i++) {
-        fs.unlinkSync(`./public/uploads/projects/${projectById.images[i]}`);
+    const isAdmin = await Admins.find({ 'token': req.headers.authorization }).populate('token');
+    if (isAdmin.length <= 0) {
+        return res.status(400).json({ statusCode: 400, message: 'Вы не авторизованны' });
     }
+
+    const projectById = await Projects.findById(id);
+    fs.unlinkSync(`./public/uploads/projects/${projectById.cover}`);
+    fs.unlinkSync(`./public/uploads/projects/${projectById.image}`);
+
 
     try {
         const deleteProject = await Projects.deleteOne({ _id: id });
 
         if (!deleteProject) {
-            return res.status(400).json({ statusCode: 400 });
+            return res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
         }
 
         res.status(200).json({ statusCode: 200, data: deleteProject });
 
     } catch (error) {
-        res.status(400).json({ statusCode: 400 });
+        res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
     }
 })
