+++ pages/api/socials/[id].js
@@ -1,5 +1,6 @@
 import dbConnect from '../../../utils/dbConnect';
 import Socials from '../../../models/Socials';
+import Admins from '../../../models/Admins';
 
 dbConnect();
 
@@ -27,39 +28,47 @@ export default async (req, res) => {
 
         case 'PUT':
             try {
+                const isAdmin = await Admins.find({ 'token': req.headers.authorization }).populate('token');
+                if (isAdmin.length <= 0)
+                    return res.status(400).json({ statusCode: 400, message: 'Вы не авторизованны' });
+
                 const social = await Socials.findByIdAndUpdate(id, req.body, {
                     new: true,
                     runValidators: true,
                 })
 
                 if (!social) {
-                    return res.status(400).json({ statusCode: 400 });
+                    return res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
                 }
 
                 res.status(200).json({ statusCode: 200, data: social });
 
             } catch (error) {
-                res.status(400).json({ statusCode: 400 });
+                res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
             }
             break;
 
         case 'DELETE':
             try {
+                const isAdmin = await Admins.find({ 'token': req.headers.authorization }).populate('token');
+                if (isAdmin.length <= 0)
+                    return res.status(400).json({ statusCode: 400, message: 'Вы не авторизованны' });
+
                 const deleteSocial = await Socials.deleteOne({ _id: id });
 
                 if (!deleteSocial) {
-                    return res.status(400).json({ statusCode: 400 });
+                    return res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
                 }
 
                 res.status(200).json({ statusCode: 200, data: deleteSocial });
 
             } catch (error) {
-                res.status(400).json({ statusCode: 400 });
+                res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
             }
             break;
 
         default:
-            res.status(400).json({ statusCode: 400 });
+            res.status(400).json({ statusCode: 400, message: 'Что то пошло не так...' });
             break;
     }
 }
